Created get_latest_prices method and created an ignored data folder
Made it so that the data object is passed from the get_latest_prices
method so that any method could just get a list of the latest prices at
any point. I also added a key called "error" so that you could check
whether an error occurred.

I also changed log_price() to use this new method to build the lines
that it writes to the file.

Then I created a folder called "data" that all of our files can be
placed. I added this to the .gitignore list so that it will never be
versioned.